requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.7.1
Flask-Cors==3.0.10
Werkzeug==2.3.7
tensorflow==2.14.0
pillow==10.3.0
python-magic-bin==0.4.14
requests==2.31.0
boto3==1.28.77
python-dotenv==1.0.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.95.0
gunicorn==20.1.0
